% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monNameGen.r
\name{monNameGen}
\alias{monNameGen}
\title{Monster Name Generator}
\usage{
monNameGen(n = 1, filters = list())
}
\arguments{
\item{n}{Number of monster names to return.}

\item{filters}{A named list with regular expressions to filter out
possibilities. The list must contain a combination of zero or more of the
following character vectors with valid regular expressions: \code{prefix},
\code{suffix}, \code{appelation}.}
}
\description{
Create a random monster name in the style of Diablo II. Stock for potential
permutations provided by
\url{http://diablo.gamepedia.com/Unique_Monsters_(Diablo_II)}.
}
\examples{
monNameGen(10)
monNameGen(10, list(prefix= "Star", appelation = "Shade"))
monNameGen(10, list(suffix = "^Kill"))
}
\author{
Derek Nedveck and Paul Egeler
}
